requirements.txt
Flask==2.3.2
Flask-Bcrypt==1.0.1
Flask-Login==0.6.2
requests==2.31.0
Flask-WTF==1.1.1
click==8.1.3
email_validator==2.0.0.post2
flask_behind_proxy==0.1.1
flask_sqlalchemy==3.0.5
GitPython==3.1.31
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
pycodestyle==2.10.0
pytest==7.4.0
Werkzeug==2.3.6
urllib3==2.0.3
flask_migrate==4.0.4
python-dotenv==1.0.0
twilio==7.15.0
openai==0.28
gunicorn==20.0.4